Refactor package structure for better organization
