+++ src/main/java/lechatbot/LeChatBotException.java
@@ -1,3 +1,5 @@
+package lechatbot;
+
 public class LeChatBotException extends Exception {
     public LeChatBotException(LeChatBot.ErrorType errorType) {
         super(errorType.getMessage());
