+++ src/main/java/lechatbot/Parser.java
@@ -1,3 +1,8 @@
+package lechatbot;
+
+import lechatbot.command.*;
+import lechatbot.task.Todo;
+
 public class Parser {
     public static Command parse(String userInput) throws LeChatBotException {
         String[] parts = userInput.split(" ", 2);
