+++ src/main/java/lechatbot/Ui.java
@@ -1,3 +1,6 @@
+package lechatbot;
+
+import lechatbot.task.TaskList;
 import java.util.Scanner;
 
 public class Ui {
