+++ src/main/java/lechatbot/command/Command.java
@@ -1,3 +1,10 @@
+package lechatbot.command;
+
+import lechatbot.LeChatBotException;
+import lechatbot.Storage;
+import lechatbot.task.TaskList;
+import lechatbot.Ui;
+
 public abstract class Command {
     public abstract void execute(TaskList tasks, Ui ui, Storage storage) throws LeChatBotException;
