+++ src/main/java/lechatbot/command/DeleteCommand.java
@@ -1,9 +1,15 @@
+package lechatbot.command;
+
+import lechatbot.*;
+import lechatbot.task.Task;
+import lechatbot.task.TaskList;
+
 import java.io.IOException;
 
-public class UnmarkCommand extends Command {
+public class DeleteCommand extends Command {
     private final int taskIndex;
 
-    public UnmarkCommand(int taskIndex) {
+    public DeleteCommand(int taskIndex) {
         this.taskIndex = taskIndex;
     }
 
@@ -12,10 +18,11 @@ public class UnmarkCommand extends Command {
         if (taskIndex < 0 || taskIndex >= tasks.size()) {
             throw new LeChatBotException("OOPS!!! The task number provided is invalid.");
         }
-        tasks.get(taskIndex).markAsNotDone();
+        Task removedTask = tasks.remove(taskIndex);
         ui.showLine();
-        System.out.println("OK, I've marked this task as not done yet:");
-        System.out.println("  " + tasks.get(taskIndex));
+        System.out.println("Noted. I've removed this task:");
+        System.out.println("  " + removedTask);
+        System.out.println("Now you have " + tasks.size() + " task" + (tasks.size() > 1 ? "s" : "") + " in the list.");
         ui.showLine();
         try {
             storage.save(tasks.getTasks());
@@ -23,4 +30,5 @@ public class UnmarkCommand extends Command {
             throw new LeChatBotException("OOPS!!! An error occurred while saving tasks.");
         }
     }
+
 }
