+++ src/main/java/lechatbot/command/ExitCommand.java
@@ -1,3 +1,9 @@
+package lechatbot.command;
+
+import lechatbot.Storage;
+import lechatbot.task.TaskList;
+import lechatbot.Ui;
+
 public class ExitCommand extends Command {
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) {
