+++ src/main/java/lechatbot/task/Deadline.java
@@ -1,3 +1,5 @@
+package lechatbot.task;
+
 import java.time.LocalTime;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
