+++ src/main/java/lechatbot/task/Event.java
@@ -1,3 +1,5 @@
+package lechatbot.task;
+
 import java.time.LocalTime;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
