+++ src/main/java/lechatbot/task/Task.java
@@ -1,3 +1,5 @@
+package lechatbot.task;
+
 public abstract class Task {
     protected String description;
     protected boolean isDone;
