+++ src/main/java/lechatbot/task/Todo.java
@@ -1,3 +1,5 @@
+package lechatbot.task;
+
 public class Todo extends Task {
     public Todo(String description) {
         super(description);
